// frontend/src/openapi/compassClient/model/updateDaySheetDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class UpdateDaySheetDto {
    'id'?: number;
    'date'?: Date;
    'dayReport'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        },
        {
            "name": "date",
            "baseName": "date",
            "type": "Date"
        },
        {
            "name": "dayReport",
            "baseName": "day_report",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return UpdateDaySheetDto.attributeTypeMap;
    }
}