// frontend/src/openapi/compassClient/models/WorkHourDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ParticipantDto } from './ParticipantDto';
import {
    ParticipantDtoFromJSON,
    ParticipantDtoFromJSONTyped,
    ParticipantDtoToJSON,
} from './ParticipantDto';

/**
 * 
 * @export
 * @interface WorkHourDto
 */
export interface WorkHourDto {
    /**
     * 
     * @type {number}
     * @memberof WorkHourDto
     */
    daySheetId?: number;
    /**
     * 
     * @type {Date}
     * @memberof WorkHourDto
     */
    date?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof WorkHourDto
     */
    confirmed?: boolean;
    /**
     * 
     * @type {number}
     * @memberof WorkHourDto
     */
    workHours?: number;
    /**
     * 
     * @type {ParticipantDto}
     * @memberof WorkHourDto
     */
    participant?: ParticipantDto;
}

/**
 * Check if a given object implements the WorkHourDto interface.
 */
export function instanceOfWorkHourDto(value: object): boolean {
    return true;
}

export function WorkHourDtoFromJSON(json: any): WorkHourDto {
    return WorkHourDtoFromJSONTyped(json, false);
}

export function WorkHourDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkHourDto {
    if (json == null) {
        return json;
    }
    return {
        
        'daySheetId': json['daySheetId'] == null ? undefined : json['daySheetId'],
        'date': json['date'] == null ? undefined : (new Date(json['date'])),
        'confirmed': json['confirmed'] == null ? undefined : json['confirmed'],
        'workHours': json['workHours'] == null ? undefined : json['workHours'],
        'participant': json['participant'] == null ? undefined : ParticipantDtoFromJSON(json['participant']),
    };
}

export function WorkHourDtoToJSON(value?: WorkHourDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'daySheetId': value['daySheetId'],
        'date': value['date'] == null ? undefined : ((value['date']).toISOString()),
        'confirmed': value['confirmed'],
        'workHours': value['workHours'],
        'participant': ParticipantDtoToJSON(value['participant']),
    };
}