// frontend/src/openapi/compassClient/model/workHourDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ParticipantDto } from './participantDto';

export class WorkHourDto {
    'daySheetId'?: number;
    'date'?: Date;
    'confirmed'?: boolean;
    'workHours'?: number;
    'participant'?: ParticipantDto;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "daySheetId",
            "baseName": "daySheetId",
            "type": "number"
        },
        {
            "name": "date",
            "baseName": "date",
            "type": "Date"
        },
        {
            "name": "confirmed",
            "baseName": "confirmed",
            "type": "boolean"
        },
        {
            "name": "workHours",
            "baseName": "workHours",
            "type": "number"
        },
        {
            "name": "participant",
            "baseName": "participant",
            "type": "ParticipantDto"
        }    ];

    static getAttributeTypeMap() {
        return WorkHourDto.attributeTypeMap;
    }
}